cleaup some matrix functions, remove unused functions from headers, rearrange and create new headers
