+++ inc/raytracing.h
@@ -14,14 +14,6 @@
 # define SPEC_DEFAULT 0.9
 # define SHINE_DEFAULT 200
 
-typedef struct s_tuple
-{
-	float	x;
-	float	y;
-	float	z;
-	float	w;
-}	t_tuple;
-
 typedef struct s_ray
 {
 	t_tuple	origin;
@@ -47,13 +39,6 @@ typedef struct s_env
 	t_tuple	w;
 }	t_env;
 
-//dev
-typedef struct s_proj
-{
-	t_tuple	pos;
-	t_tuple	v;
-}	t_proj;
-
 typedef struct s_i
 {
 	float	t;
@@ -73,37 +58,7 @@ typedef struct s_xs
 	float	t[MAX_OBJECTS];
 }	t_xs;
 
-typedef struct s_matrix4
-{
-	float	m[4][4];
-}	t_matrix4;
 
-typedef struct s_matrix3
-{
-	float	m[3][3];
-}	t_matrix3;
-
-typedef struct s_matrix2
-{
-	float	m[2][2];
-}	t_matrix2;
-typedef struct s_matrix_ctx
-{
-	t_matrix4	m;
-	size_t		size;
-}	t_matrix_ctx;
-
-typedef union s_matrix_union
-{
-	t_matrix2	m2;
-	t_matrix3	m3;
-}	t_matrix_union;
-
-typedef struct s_matrix_result
-{
-	t_matrix_union	m;
-	size_t			size;
-}	t_matrix_result;
 
 typedef struct s_quad
 {
@@ -158,35 +113,21 @@ typedef struct	s_calc_vars
 {
 	float	xoffset;
 	float	yoffset;
-	float	world_x;
-	float	world_y;
+	float	view_port_x;
+	float	view_port_y;
 }	t_calc_vars;
 
 //lighting.c
-t_light	init_point_light(t_tuple pos, t_color color, float ratio);
-t_tuple	reflect(t_tuple in, t_tuple normal);
-//t_color	lighting(t_material m, t_light l, t_tuple p, t_tuple e_v, t_tuple n_v);
+t_light		init_point_light(t_tuple pos, t_color color, float ratio);
+t_tuple		reflect(t_tuple in, t_tuple normal);
+t_color		lighting(t_minirt m, t_comps c, bool in_shadow, t_light_vars *lv);
 
 //normal.c
-t_tuple	normal_at(t_minirt *minirt, t_scene_obj *obj, t_tuple point);
-t_tuple	normal_at_cylinder(t_scene_obj *cylinder, t_tuple point); //testing only
-
-//tuples
-t_tuple	create_vector(float x, float y, float z);
-t_tuple	create_point(float x, float y, float z);
-
-//tuple operations
-t_tuple	addition_tuples(t_tuple tuple1, t_tuple tuple2);
-t_tuple	substraction_tuples(t_tuple tuple1, t_tuple tuple2);
-t_tuple	negate_tuple(t_tuple tuple1);
-t_tuple	scalar_multiply_tuple(t_tuple tuple1, float scalar);
-t_tuple	scalar_divide_tuple(t_tuple tuple1, float scalar);
-float	magnitude_tuple(t_tuple tuple);
-t_tuple	normalize_tuple(t_tuple tuple);
-float	dot_tuple(t_tuple tuple1, t_tuple tuple2);
-t_tuple	cross_tuple(t_tuple tuple1, t_tuple tuple2);
+t_tuple		normal_at(t_minirt *minirt, t_scene_obj *obj, t_tuple point);
+t_tuple		normal_at_cylinder(t_scene_obj *cylinder, t_tuple point);
 
 //colors
+t_color		color(float r, float g, float b);
 t_color		addition_color(t_color color1, t_color color2);
 t_color		substraction_color(t_color color1, t_color color2);
 t_color		multiply_color(t_color color1, t_color color2);
@@ -196,43 +137,12 @@ void		colour_unitrgb_rgba(t_color *c);
 void		colour_rgba_unitrgb(t_color *c);
 void		color_convert(t_scene_obj *obj);
 
-//matrix math
-t_matrix4	multiply_mtrx_by_mtrx(t_matrix4 m, t_matrix4 m2);
-t_tuple		multiply_mtrx_by_tuple(t_matrix4 m, t_tuple t1);
-t_matrix4	transpose_matrix(t_matrix4 m);
-float		determinant_matrix4(t_matrix4 m);
-float		determinant_matrix3(t_matrix3 m);
-float		determinant_matrix2(t_matrix2 m);
-t_matrix_result	sub_matrix(t_matrix_ctx *ctx, size_t row, size_t col);
-float		minor_matrix(t_matrix_ctx *ctx, int row, int col);
-float		cofactor_matrix(t_matrix_ctx *ctx, int row, int col);
-t_matrix4	inverse_matrix(t_matrix4 m, bool *success);
-bool		is_invertible_matrix4(t_matrix4 m);
-bool		is_invertible_matrix3(t_matrix3 m);
-bool		is_invertible_matrix2(t_matrix2 m);
-t_matrix4	scalar_multiply_matrix(t_matrix4 m, float scalar);
-t_matrix4	addition_matrix(t_matrix4 a, t_matrix4 b);
-
-//matrix operators
-t_matrix4	identity(void);
-t_matrix4	translation(float x, float y, float z);
-t_matrix4	scaling(float x, float y, float z);
-t_matrix4	rotation_x(float theta);
-t_matrix4	rotation_y(float theta);
-t_matrix4	rotation_z(float theta);
-
-//matrix utils
-void	matrix_fill_zero(t_matrix4 *m);
-
 //rays
-t_ray	create_ray(t_tuple vector, t_tuple point);
-t_tuple	position_ray(t_ray ray, float t);
-t_i		hit(t_xs *xs, t_scene_obj *obj_from);
-t_ray	transform(t_ray r, t_matrix4 m);
-void	set_transform(t_scene_obj *obj, t_matrix4 m);
-t_i		intersection(float intersection, void *obj);
-void	intersections(t_xs	*xs, t_i i1, t_i i2);
-t_xs	intersect_world(t_minirt *minirt, t_ray r);
+t_ray		create_ray(t_tuple vector, t_tuple point);
+t_tuple		position_ray(t_ray ray, float t);
+t_i			hit(t_xs *xs, t_scene_obj *obj_from);
+t_ray		transform(t_ray r, t_matrix4 m);
+t_xs		intersect_world(t_minirt *minirt, t_ray r);
 
 //shading
 t_comps		prepare_computations(t_minirt *minirt, t_i i, t_ray r);
@@ -243,15 +153,16 @@ bool		is_shadowed(t_minirt *minirt, t_tuple point, t_scene_obj *obj);
 //camera
 t_matrix4	view_transform(t_tuple from, t_tuple to, t_tuple up);
 void		init_camera(t_minirt *minirt);
-t_ray		ray_for_pixel(t_minirt *minirt, t_camera *c, int px, int py);
 
 //intersections.c
 t_xs		*intersects_sphere(t_scene_obj *obj, t_ray r, t_xs *xs);
 t_xs		*intersect(t_minirt *minirt, t_scene_obj *obj, t_ray ray, t_xs *xs);
+t_i			intersection(float intersection, void *obj);
+void		intersections(t_xs	*xs, t_i i1, t_i i2);
 
 //intersect_cylinder.c
-t_xs	*intersects_cylinder(t_scene_obj *obj, t_ray r, t_xs *xs);
-void	init_i_to_zeroes(t_i *i1, t_i *i2);
+t_xs		*intersects_cylinder(t_scene_obj *obj, t_ray r, t_xs *xs);
+void		init_i_to_zeroes(t_i *i1, t_i *i2);
 
 //transformation functions
 t_matrix4	generate_transformation_mtrx(t_minirt *minirt, t_scene_obj *obj);
@@ -261,52 +172,10 @@ t_matrix4	plane_rotation(t_scene_obj *obj);
 
 //utils
 bool		is_equal(float a, float b);
-uint32_t	calculate_hit(t_minirt *minirt, size_t x, size_t y);
-
-//print utils
-void	print_ray(t_ray r);
-void	print_xs(t_xs *xs);
-void	print_matrix(t_matrix4 m, char *msg, int size);
-void	print_tuple(t_tuple t);
-void	print_colour(t_color c);
-void	print_comps(t_comps *comps);
-void	print_camera(t_camera *cam);
-void	fun_test_parsed_output(char **av, t_parse *ps);
 
 //objects.c
 t_scene_obj	sphere(t_minirt *minirt);
 t_scene_obj	plane(t_minirt *minirt);
 t_scene_obj	cylinder(t_minirt *minirt);
-
-//dev
-float	**create_matrix(size_t size, int flag);
-void	unit_tests_3x3(void);
-void	unit_tests_4x4(void);
-void	unit_tests_transform_matrices(void);
-void	test_normal_at_sphere(t_minirt *m);
-void	test_reflect(t_minirt m);
-void	test_reflect_extra(t_minirt m);
-void	test_point_light_material(void);
-void	test_point_light_reflections(void);
-void	test_prepare_computations_outside(t_minirt *minirt, char **av);
-void	test_prepare_computations_inside(t_minirt *minirt, char **av);
-void	test_shading_an_intersection(t_minirt *minirt, char **av);
-void	test_shading_an_intersection_from_inside(t_minirt *minirt, char **av);
-void	test_ray_misses_obj(t_minirt *minirt);
-void	test_ray_hits_obj(t_minirt *minirt);
-void	test_intersection_behind_ray(t_minirt *minirt);
-void	test_orientation(void);
-void	test_camera(t_minirt *minirt);
-void	test_ray_for_pixel(t_minirt *minirt);
-void	test_intersect_two_spheres(t_minirt *minirt);
-void	print_colour(t_color c);
-void	test_shape(t_minirt *minirt);
-void	test_intersect_generic(void);
-void	test_plane_normal(t_minirt *minirt);
-void	test_plane_intersect(t_minirt *minirt);
-void	test_render_world(t_minirt *minirt);
-void	test_cylinder_rotation(void);
-void	test_cylinder(t_minirt *minirt);
-void	test_cylinder2(t_minirt *minirt);
 
 #endif
