+++ scenes/ambient_test1.rt
@@ -1,0 +1,5 @@
+A 0.3 255,0,0
+C 0,0,-3 0,0,1  60
+L 0,0,-3 1
+
+sp 0,0,0 2 255,255,255
